Fix SignUpForm export for Next.js page
